backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
pydantic[email]==2.7.1
scikit-learn==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
razorpay==1.4.2
openai==1.68.2
bcrypt==4.0.1